refactor: installed Prettier. added *components* folder and moved all To Do List related componenets in
