+++ src/App.css
@@ -1,15 +1,15 @@
-@import url('https://fonts.googleapis.com/css2?family=Poppins&display=swap');
+@import url("https://fonts.googleapis.com/css2?family=Poppins&display=swap");
 
 * {
-  font-family: 'Poppins', sans-serif;
+  font-family: "Poppins", sans-serif;
   margin: 0;
   padding: 0;
   box-sizing: border-box;
-  color: #6F9283;
+  color: #6f9283;
 }
 
 body {
-  background: #CDC6A5;
+  background: #cdc6a5;
   display: flex;
   justify-content: center;
   align-items: center;
@@ -26,7 +26,7 @@ h1 {
 }
 
 .TodoWrapper {
-  background: #F0DCCA;
+  background: #f0dcca;
   margin-top: 5rem;
   padding: 2rem;
   border-radius: 5px;
@@ -40,19 +40,19 @@ h1 {
 .todo-input {
   outline: none;
   background: none;
-  border: 1px solid #6F9283;
+  border: 1px solid #6f9283;
   padding: 0.5rem 1rem;
   margin-top: 1rem;
   margin-bottom: 2rem;
-  width: 300px; 
+  width: 300px;
 }
 
 .todo-input::placeholder {
-  color: #6F92834d;
+  color: #6f92834d;
 }
 
 .todo-btn {
-  background: #8D9F87;
+  background: #8d9f87;
   margin-left: 0.25rem;
   color: #fff;
   padding: 0.5rem;
@@ -62,7 +62,7 @@ h1 {
 }
 
 .todo-btn:hover {
-  border: 1px solid #6F92834d;
+  border: 1px solid #6f92834d;
 }
 
 .Todo {
