+++ src/App.tsx
@@ -1,12 +1,12 @@
-import './App.css'
-import TodoWrapper from './TodoWrapper'
+import "./App.css";
+import TodoWrapper from "./components/TodoWrapper";
 
 function App() {
   return (
     <>
       <TodoWrapper />
     </>
-  )
+  );
 }
 
-export default App
+export default App;
