--- src/TodoForm.tsx
@@ -1,11 +1,0 @@
-function TodoForm () {
-
-  return (
-    <form className="TodoForm">
-        <input type="text" className="todo-input" placeholder="What is the task today?" onChange={ (e) => {console.log(e.target.value)}}/>
-        <button type="submit" className="todo-btn">Add Task</button>
-    </form>
-  )
-}
-
-export default TodoForm;
