+++ src/components/TodoWrapper.tsx
@@ -2,14 +2,14 @@ import EditTodoForm from "./EditTodoForm";
 import Todo from "./Todo";
 import TodoForm from "./TodoForm";
 
-function TodoWrapper () {
+function TodoWrapper() {
   return (
     <div className="TodoWrapper">
       <TodoForm />
       <Todo />
       <EditTodoForm />
     </div>
-  )
+  );
 }
 
 export default TodoWrapper;
